Move validation files into new folder
Related to #230
