+++ src/app/utils/validations/index.ts
@@ -1,0 +1,7 @@
+import { isValidProjectData } from "./projectValidation";
+import { isValidReferralData } from "./referralValidation";
+
+export {
+  isValidProjectData,
+  isValidReferralData
+}
